refactor(game): move out Image implementation into separate file
