+++ src/game/CMakeLists.txt
@@ -36,6 +36,7 @@ if(${LVF_GRAPHICS_BACKEND} STREQUAL SDL2)
     set(LVF_GRAPHICS_BACKEND_SRC
         graphics/SDL2/common_SDL2.cpp
         graphics/SDL2/graph_line_SDL2.cpp
+        graphics/SDL2/image_SDL2.cpp
         graphics/SDL2/input_SDL2.cpp
         graphics/SDL2/ttf_SDL2.cpp
         graphics/SDL2/window_SDL2.cpp
